🚚 [RUMF-951] rename rumRecorderPublicApi to recorderApi
