+++ packages/rum/src/boot/recorderApi.ts
