+++ packages/rum/src/boot/rum.entry.ts
@@ -2,7 +2,7 @@ import { defineGlobal, getGlobalObject } from '@datadog/browser-core'
 import { makeRumPublicApi, RumPublicApi, startRum } from '@datadog/browser-rum-core'
 
 import { startRecording } from './startRecording'
-import { makeRecorderApi } from './rumRecorderPublicApi'
+import { makeRecorderApi } from './recorderApi'
 
 const recorderApi = makeRecorderApi(startRecording)
 export const datadogRum = makeRumPublicApi(startRum, recorderApi)
